APIs/requirements.txt
bokeh==2.4.3
Django==3.2.8
django_admin_tools==0.9.2
django_filter==21.1
django_pandas==0.6.6
folium==0.12.1.post1
geemap==0.16.7
geojson==2.5.0
geopandas==0.11.1
netCDF4==1.6.3
numpy==1.23.3
pandas==1.5.0
plotly==5.8.0
PyMySQL==1.0.2
regionmask==0.9.0
rioxarray==0.3.1
Shapely==1.8.4
SQLAlchemy==1.4.22
sshtunnel==0.4.0
xarray==2022.3.0
